Apply the same DNS resolve strategy to wireguard service
